services/defect-prediction-service/requirements.txt
fastapi
uvicorn
openai
numpy
pandas
scikit-learn
tensorflow
networkx
pydantic
python-dotenv
matplotlib
seaborn
GitPython